+++ packages/core/src/extensions-shared/README.md
@@ -1,0 +1,3 @@
+### @blocknote/core/src/extensions-shared
+
+Helper functions / base plugins for @blocknote/core/src/extensions
