+++ packages/core/src/extensions-shared/suggestion/SuggestionItem.ts
