+++ packages/core/src/extensions-shared/suggestion/SuggestionPlugin.ts
@@ -1,10 +1,10 @@
 import { findParentNode } from "@tiptap/core";
 import { EditorState, Plugin, PluginKey } from "prosemirror-state";
 import { Decoration, DecorationSet, EditorView } from "prosemirror-view";
-import type { BlockNoteEditor } from "../../../editor/BlockNoteEditor";
-import { BlockSchema } from "../../../schema/blocks/types";
-import { InlineContentSchema } from "../../../schema/inlineContent/types";
-import { StyleSchema } from "../../../schema/styles/types";
+import type { BlockNoteEditor } from "../../editor/BlockNoteEditor";
+import { BlockSchema } from "../../schema/blocks/types";
+import { InlineContentSchema } from "../../schema/inlineContent/types";
+import { StyleSchema } from "../../schema/styles/types";
 import { BaseUiElementState } from "../BaseUiElementTypes";
 import { SuggestionItem } from "./SuggestionItem";
