+++ packages/core/src/extensions/FormattingToolbar/FormattingToolbarPlugin.ts
@@ -2,7 +2,7 @@ import { isNodeSelection, posToDOMRect } from "@tiptap/core";
 import { EditorState, Plugin, PluginKey } from "prosemirror-state";
 import { EditorView } from "prosemirror-view";
 
-import { EventEmitter } from "../../@util/EventEmitter";
+import { EventEmitter } from "../../util/EventEmitter";
 import type { BlockNoteEditor } from "../../editor/BlockNoteEditor";
 import { BlockSchema } from "../../schema/blocks/types";
 import { InlineContentSchema } from "../../schema/inlineContent/types";
@@ -10,7 +10,7 @@ import { StyleSchema } from "../../schema/styles/types";
 import {
   BaseUiElementCallbacks,
   BaseUiElementState,
-} from "../@shared/BaseUiElementTypes";
+} from "../../extensions-shared/BaseUiElementTypes";
 
 export type FormattingToolbarCallbacks = BaseUiElementCallbacks;
