+++ packages/core/src/extensions/TableHandles/TableHandlesPlugin.ts
@@ -1,6 +1,6 @@
 import { Plugin, PluginKey, PluginView } from "prosemirror-state";
 import { Decoration, DecorationSet, EditorView } from "prosemirror-view";
-import { EventEmitter } from "../../@util/EventEmitter";
+import { EventEmitter } from "../../util/EventEmitter";
 import { nodeToBlock } from "../../api/nodeConversions/nodeConversions";
 import { DefaultBlockSchema } from "../../blocks/defaultBlocks";
 import type { BlockNoteEditor } from "../../editor/BlockNoteEditor";
