+++ packages/core/src/util/browser.ts
