+++ packages/core/src/util/typescript.ts
