+++ packages/react/src/components-shared/ColorPicker/ColorIcon.tsx
