+++ packages/react/src/components/FormattingToolbar/DefaultButtons/CreateLinkButton.tsx
@@ -2,8 +2,8 @@ import { useCallback, useMemo, useState } from "react";
 import { BlockNoteEditor, BlockSchema } from "@blocknote/core";
 import { RiLink } from "react-icons/ri";
 
-import { ToolbarInputDropdownButton } from "../../@shared/Toolbar/ToolbarInputDropdownButton";
-import { ToolbarButton } from "../../@shared/Toolbar/ToolbarButton";
+import { ToolbarInputDropdownButton } from "../../../components-shared/Toolbar/ToolbarInputDropdownButton";
+import { ToolbarButton } from "../../../components-shared/Toolbar/ToolbarButton";
 import { EditHyperlinkMenu } from "../../HyperlinkToolbar/EditHyperlinkMenu/components/EditHyperlinkMenu";
 import { useSelectedBlocks } from "../../../hooks/useSelectedBlocks";
 import { useEditorChange } from "../../../hooks/useEditorChange";
