+++ packages/react/src/components/HyperlinkToolbar/EditHyperlinkMenu/components/EditHyperlinkMenu.tsx
@@ -8,8 +8,8 @@ import {
   useState,
 } from "react";
 import { RiLink, RiText } from "react-icons/ri";
-import { ToolbarInputDropdown } from "../../../@shared/Toolbar/ToolbarInputDropdown";
-import { ToolbarInputDropdownItem } from "../../../@shared/Toolbar/ToolbarInputDropdownItem";
+import { ToolbarInputDropdown } from "../../../../components-shared/Toolbar/ToolbarInputDropdown";
+import { ToolbarInputDropdownItem } from "../../../../components-shared/Toolbar/ToolbarInputDropdownItem";
 
 export type EditHyperlinkMenuProps = {
   url: string;
